+++ src/mainwindow/roomeditattrdlg.cpp
@@ -1109,7 +1109,7 @@ void RoomEditAttrDlg::terrainToolButtonToggled(bool val)
 
     const RoomTerrainEnum rtt = [this]() -> RoomTerrainEnum {
         // returns the first one that's checked, or UNDEFINED.
-        for (uint i = 0; i < NUM_ROOM_TERRAIN_TYPES; ++i) {
+        for (size_t i = 0; i < NUM_ROOM_TERRAIN_TYPES; ++i) {
             const auto rtt = static_cast<RoomTerrainEnum>(i);
             if (const QToolButton *const ptr = roomTerrainButtons[rtt])
                 if (ptr->isChecked())
